added entities actions & filters
