+++ dapp/src/scss/components/dashboard/EntityDirectory.scss
@@ -28,11 +28,13 @@
     width: 100%;
     margin-bottom: 2em;
     padding: 0 50px;
+    justify-content: space-between;
 
     @include element(filter) {
       position: relative;
       cursor: pointer;
       z-index: 1;
+      width: 20%;
 
       @include element(icon) {
         display: inline-block;
@@ -74,7 +76,6 @@
         display: none;
         @include flex-align(left, middle);
         width: 188px;
-        height: 138px;      
         border-radius: 5px;
         box-shadow: 0 0 8px 0 rgba(0, 0, 0, 0.3);
         background-color: #ffffff;
@@ -157,7 +158,7 @@
     @include element(buttons) {
       @include xy-grid();
       @include flex-align(center, middle);
-      margin: 0 auto 0 24%;
+      // margin: 0 auto;
 
       > button {
         &:last-child {
@@ -194,12 +195,15 @@
       font-size: 11px;
       @include xy-grid();
       @include flex-align(null, middle);
+      width: 20%;
+
       > span {
         width: 20px;
         height: 20px;
         display: inline-block;
         cursor: pointer;
         margin-right: 5px;
+        margin-left: auto;
 
         .fa {
           font-size: 20px;
@@ -264,7 +268,6 @@
   }
 
   @include element(loader) {
-    margin-top: 40px;
     text-align: center;
     
     @include block(loader) {
