+++ dapp/src/selectors/directory.js
@@ -1,4 +1,8 @@
 import { createSelector } from 'reselect'
+import { getAccountAddress } from './accounts'
+import sortBy from 'lodash/sortBy'
+
+export const getCommunityAddress = state => state.screens.directory.communityAddress
 
 export const getEntities = createSelector(
   state => state.screens.directory.listHashes,
@@ -9,11 +13,19 @@ export const getEntities = createSelector(
 export const getUsersEntities = createSelector(
   state => state.screens.directory.usersResults,
   state => state.entities.communityEntities,
-  (usersResults, communityEntities) => usersResults.map(account => communityEntities[account]).filter(obj => !!obj)
+  (usersResults, communityEntities) => sortBy(usersResults.map(account => communityEntities[account]).filter(obj => !!obj), ['updatedAt']).reverse() || []
 )
 
 export const getBusinessesEntities = createSelector(
   state => state.screens.directory.merchantsResults,
   state => state.entities.communityEntities,
-  (merchantsResults, communityEntities) => merchantsResults.map(account => communityEntities[account]).filter(obj => !!obj)
+  (merchantsResults, communityEntities) => sortBy(merchantsResults.map(account => communityEntities[account]).filter(obj => !!obj), ['updatedAt']).reverse() || []
+)
+
+export const checkIsAdmin = createSelector(
+  getAccountAddress,
+  state => state.entities.communityEntities,
+  (accountAddress, communityEntities) => {
+    return (communityEntities[accountAddress] && communityEntities[accountAddress].isAdmin) || false
+  }
 )
