+++ dapp/src/services/api/entities.js
@@ -2,15 +2,15 @@ import request from 'superagent'
 
 export const fetchCommunityEntities = (apiRoot, { communityAddress, entityType }) => {
   const path = entityType
-    ? `${apiRoot}/entities?communityAddress=${communityAddress}&type=${entityType}`
-    : `${apiRoot}/entities?communityAddress=${communityAddress}`
+    ? `${apiRoot}/entities/${communityAddress}?type=${entityType}`
+    : `${apiRoot}/entities/${communityAddress}`
   return request.get(path).then(response => response.body)
 }
 
-export const createEntitiesMetadata = (apiRoot, { accountId, metadata }) =>
-  request.put(`${apiRoot}/entities/${accountId}`)
+export const createEntitiesMetadata = (apiRoot, { communityAddress, accountId, metadata }) =>
+  request.put(`${apiRoot}/entities/${communityAddress}/${accountId}`)
     .send({ metadata })
     .then(response => response.body)
 
-export const fetchEntity = (apiRoot, { account }) =>
-  request.get(`${apiRoot}/entities/${account}`).then(response => response.body)
+export const fetchEntity = (apiRoot, { communityAddress, account }) =>
+  request.get(`${apiRoot}/entities/${communityAddress}/${account}`).then(response => response.body)
